% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accident_2014}
\alias{accident_2014}
\title{A data frame containing car accidents for US in 2014}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 30056 rows and 50 columns.}
\source{
US National Highway Traffic Safety Administration's
}
\usage{
accident_2014
}
\description{
A data frame containing car accidents for US in 2014
}
\examples{

\dontrun{accident_2014}
}
\keyword{datasets}
