% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accident_2015}
\alias{accident_2015}
\title{A data frame containing car accidents for US in 2015}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 32166 rows and 52 columns.}
\source{
US National Highway Traffic Safety Administration's
}
\usage{
accident_2015
}
\description{
A data frame containing car accidents for US in 2015
}
\examples{

\dontrun{accident_2015}
}
\keyword{datasets}
