% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Create a graphical map with the dataset}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{It is the input of the function, consisting of ideally integers in order to specify correct files
and subsets.
If a string is entered, an error can appear}

\item{year}{It is the input of the function, consisting of ideally integers in order to specify correct files
and subsets.
If a string is entered, an error can appear}
}
\value{
Finally, a plot is retrieved
}
\description{
This function draws a map of USA and places in it certain points
defined in the dataset. In order to accomplish this task, it is
created a function which takes two arguments as inputs, they
should be numeric values otherwise an error can be retrieved
(year and sate_number). First, gathers a given dataset thanks
to a number provided as years and then filters the dataset
according to a state number provided (examining if the state
is in the dataset and if there is data for the given state,
a basic error handling procedure). Then, it cleans the column
longitude, deleting values which are higher than 900 and column
latitude, this time if values are higher than 90. Finally, a map
is produced thanks to the package \code{maps} and points specified by
the dataset are placed in it thanks to the package \code{graphics}
and the function \code{points}.
}
\examples{
\dontrun{fars_map_state()}

}
