% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Data List By Year}
\usage{
fars_read_years(years)
}
\value{
A subset of a dataframe (final output) is retrieved.
}
\description{
This function retrieves a list consisting of datasets.
It uses the previously described functions (\code{make_filename}, \code{fars_read})
while it is inserted an error handler. This time it allows inputting
a vector thanks to the function \code{lapply} (for each element of the vector,
applies the specified function). First, thanks to the functions
\code{make_filename} and \code{fars_read}; gathers a dataset (expecting that the file
will be in the working directory) and then the function reads it and tries
to select a given subset through the external package \code{dplyr}, adding
the variable year to the dataset.
}
\note{
In case of error, no data is retrieved
and the message invalid year is displayed.
}
