% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summary of the dataset}
\usage{
fars_summarize_years(years)
}
\value{
A summary of a dataframe as a table is retrieved
}
\description{
This function first uses the already described one, \code{fars_read_years}.
Using the external package \code{dplyr}, first bind the provided list into
a compact dataset with the function \code{bind_rows}. And creates a summary
dataset thanks to the function \code{summarize}, using the function \code{n()} which
counts the number of occurrences stratified by month and year thanks
to the function \code{group_by}. Finally, it is created a contingency table
thanks to the package \code{tidyr}, which displays in a better the summary
data (number of observations by year and moth)
}
