% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{make_filename}
\alias{make_filename}
\title{Creates a file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{It is the input of the function, consisting of an ideally integer in order to specify a correct file.
If a string is entered, an error can appear.}
}
\value{
the function retrieves a long string indicating the name
of a file being the integer inserted inside of the string.
}
\description{
The function \code{make_filename} takes as argument a number
(it should be a number otherwise an error could appear), it is converted to
integer.
}
\note{
In conjunction
with the previous function it can accomplish in a handy way the import of
datasets, just entering a given number as input.
}
\examples{
\dontrun{make_filename()}

}
