% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\docType{data}
\name{year}
\alias{year}
\title{Creates a file name}
\format{An object of class \code{numeric} of length 1.}
\usage{
year
}
\value{
the function retrieves a long string indicating the name
of a file being the integer inserted inside of the string.
}
\description{
The function \code{make_filename} takes as argument a number
(it should be a number otherwise an error could appear), it is converted to
integer.
}
\note{
In conjunction
with the previous function it can accomplish in a handy way the import of
datasets, just entering a given number as input.
}
\keyword{datasets}
