% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{fars_read}
\alias{fars_read}
\title{Reads data and output a dataframe}
\usage{
fars_read(filename)
}
\value{
And then, it is converted as a dataframe (final output) through the function
}
\description{
Reads data and output a dataframe
}
